Clarify re-registration docs to explain that unregistered experimental image
The experimental image is simply copied to a 3D format during initial registration with undergoing any movement itself. This copy can be registered to the atlas image using `register_reg`, where the "re"-registration is merely finding the copied experimental file with a registration suffix. Clarify this application to show that essentially original experimental files can be initially registered to an atlas using this method.