Raise exception if metadata could not be updated successfully
Reading image5d and metadata with metadata update turned on will fail if image5d is not available since image5d is required for the update. Rather than allowing the metadata load to be incomplete and to provide better feedback when metadata could not be loaded, raise an exception when image5d could not be loaded and the metadata version is below that of the current version.