Invert 3D visualization along the z-axis
When matching the xy orientation of 2D Matplotlib images with 3D visualizations in Mayavi, the z-axis increases downward rather than upward. To change this "handedness," provide and option to flip visualized data along the z-axis, controlled by a `Visualization` attribute. Also flip the orientation axes, simply turning of the arrowheads since the cones to avoid an inverted cone relative to the repositioned axis.