Option to further group bar groups with vertical spans
Provide parameters to show vertical spans that shade the background of bar groups in alternating gray and white to group sets of bar groups. The parameter points to a data frame column where changes in values specify a new group. Also provide an option to labels spans based on the values in this column. Apply the vertical spans when generate bar plots for volume stats from effect sizes.