Option to reverse the order of calculated label differences
Similar to the option provided when calculating R stats in commit 23c9391, provide a parameter to reverse the order of conditions when mapping differences by condition to a labels map. Turn this option off by default in the library function, but turn it on when calling it from the labels difference image generation function in `register`.