Fix saving labels image in `AtlasEditor`
Saving atlas labels failed if the given path had a different extension from that of the labels image. Since the labels image has already been loaded after checking for a number of potential extensions, use the path of the loaded image. Change `register.read_sitk` to return the full path of the found image rather than only the extension.