Option to treat histology regions corresponding to given labels as foreground for DSC
Some labeled regions corresponding to empty space in histology images, such as ventricles. These regions are treated as background when thresholding for the DSC overlap measure, incorrectly penalizing this labeling. As a workaround, provide a profile option to treat histology areas corresponding to given labels as foreground. For simplicity, labels above a given value (or below their negative value) are all treated as foreground to avoid requiring the labels reference to be loaded. Apply to E11.5 atlas profile since it has large labeled ventricles.