Specify label range when adding labels for overlap measurement
One label in the ABA ontology has a higher ID than those of ventricles, meaning that adding ventricular area is not as simple as taking all labels above a given number. To avoid including this label, change the label specification to take a range instead. Also fix the lower bounds for E11.5 to avoid taking other non-ventricular labels.