Update ROI to stay within bounds before 3D display
Previously 3D points display in the ROI GUI resized the ROI if it extended beyond bounds but did not check to ensure that the offset itself stayed within bounds. To keep the ROI within bounds, check and if necessary shift the offset before adjusting the ROI size. Also fix feedback to come after resetting segments, which clears feedback, although the feedback may still not appear presumably if the GUI has not yet been fully loaded.