Decrease verification tolerance for verifications with isotropy
The new verification method using the Hungarian algorithm uses `distance.cdist` to measure point distances, which appears to assume isotropy. Now that the blobs are scaled to isotropy for this approach, the verification tolerances need to be adjusted for isotropy. Decrease tolerance sizes, especially in the z-axis since it will be scaled later, but also in the x/y to decrease the hypotenuse of the tolerance used as the distance threshold for matches.