Fix volumes stats for unmirrored atlases
Eroding images for interior-border metrics has assumed that atlases are symmetric, only working on half of them and mirroring to the other half. To erode atlases that have not been mirrored, use the "extend_atlas" and "labels_mirror" register profile settings to check whether the atlas has been set to be mirrored in addition to checking the `atlas` parameter. This change fixes stats for unmirrored atlases when a mirrored eroded image size would have been off by 1 and avoids mirroring the interior-border region masks.