Move SimpleITK install from Conda to Pip
SimpleITK from Conda requires a separate channel that can conflict with the current channel setup and require many dependency changes with installation or removal. Although Conda currently gives a slightly newer release of SimpleITK, the recommended pathway would be to replace it with SimpleElastix anyway, so the PyPI version should be sufficient for now.

For simplicity and completeness, also move the SimpleITK installation into the `.yml` files since at least some version of SimpleITK is highly recommended for Clrbrain. With SimpleITK in these configuration files, users can create an environment with them to obtain an essentially complete environment for Clrbrain and use the Bash scripts only for finer-tuned installations.