Change dependency checking to warnings rather than exit conditions
Warn rather than exit when dependencies are not met since these dependencies may not be required for installation. Many installations from Git repos or even through Pip required compilation (eg Traits, Python-Javabridge), but many of these packages from Conda include binaries and do not require compilation. Python-Javabridge may still fail because of its current triple dependence on Java, Git, and gcc, but the packages is now optional for Clrbrain. The dependencies are still checked to warn the user, but no longer exit since the failed installations should still allow Clrbrain to run. Since the command-line flag to not check dependencies was to avoid this exit, the flag is no longer required, which also has the side benefit of providing warnings for all runs.