Customize the main group comparison in R stats
The column specifying the main group to compare has been assumed to be "Geno" except for means-based stats, which uses "Condition" instead. To customize this column, add a parameter for the main group in `statsByRegion`. Since jitter plots require both a main and sub-group column, manually revert to "Geno" as the main group when comparing "Condition".