Simplify erosion to keep only if above the target fraction
Label erosion has reduced the threshold target fraction to 1% of the original label size for the minimum filter size to aggressively erode unless the label would be lost. Now that a minimum filter size above 1 can be specified, this exception is less useful since it assumes that the filter size has been maximally reduced. Since the goal of specifying a higher minimum filter size is to erode more conservatively, this exception can simply be removed. This change simplifies erosion so that it will only take place if the label is big enough to remain above the target fraction in all cases.

- Remove target size fraction exception for smallest allowed filter size
- Use the original region size instead of recalculating