Exclude and readd labels from labels rather than markers image during watershed reannotation
When excluding labels during watershed reannotation, exclude and readd labels from the labels image itself rather than from the markers image since in theory the markers could have been modified. These labels must still be removed from the markers image as a workaround for the watershed mask bug. This change makes no difference for the current implementation, where the excluded labels are those that were not eroded in the markers.