Replace closing filter with in-painting for lateral edge extension
A closing filter had been used to fill small holes such as ventricles but only applied to the first plane to avoid artifacts and still caused artifacts such as losing thin sections. Replace the closing filter with in-painting based on the thresholded reference (eg histology) image, which both fills in these holes and also fills other unlabeled areas. Note that the profiles will need to be updated with this dictionary entry.