Reuse non-in-painted template for subsequent lateral extension rounds
When the in-painted template is used for subsequent rounds of extension, the labels may accumulate pixels that are no longer relevant in succeeding planes since pixels are only added and not subtracted. Using the non-in-painted template errs on the side of fewer pixels, which will be in-painted based on the current plane anyway.