Option for edge-aware reannotation in lateral edge extension
Use the edge-aware reannotation during lateral edge extension. This reannotation is a 2D/3D implementation, where the edge map is in 3D, but the reannotation occurs in serial 2D, with each generated plane becoming the template for the next plane to give smooth transitions from plane to plane. During the erosion step, the labels are allowed to disappear to emulate their tapering off.