Build discrete colormaps from an evenly spaced grid
Using random numbers to build a discrete colormap had the side effect of allowing colors to cluster or even duplicate one another. To ensure a given spacing between colors, use a mesh grid instead, where the grid combinations become the RGB values. Titrate the spacing to maximize it for the number of colors requested.