Store image5d shapes from metadata file
Shape data stored as metadata has been ignored on the assumption that image5d would itself contain this information, but the full image5d array may not be available such as for large image files. To recover this shape information without the full array, store the shape information as a `config` module variable. For consistency, save the full 2D array, but eventually only a single shape may be saved since image5d is assumed to be a single image.