Change largest cluster metric from fraction to number of nuclei
For consistency with noise measurement, simply record the number of nuclei rather than the fraction of clustered nuclei within the largest cluster. This change avoids the need to calculate a weighted mean for super-regions, and the fraction can be calculated based on the total number of nuclei minus noise nuclei.