Fix k-nearest-neighbor plot overlay
The overlay did not update the plot when the figure was opened and closed before the plot update, and the overlay did not distinguish lines from different samples. To take advantage of plotting groups in `plot_2d.plot_lines`, combine the data frames from each sample to generate a single plot.

- Return the data frames at different zoom levels rather than the plot axes from k-nearest-neighbors
- Plot overlaid points of all samples at each zoom level
- Parameter to change the plot legend name
- Only show combined rather than individual sample plots when calculating for multiple samples