Fallback to match moving image resolution to that of fixed image during registration
Images may fail to register because of inadequate overlap errors, even if the images are of similar shape and only differ in resolution. As a fallback, try matching the moving image's spacing to that of the fixed image and re-running the registration, especially since the final output of registered files has this same resolution anyway. The spacing is not applied automatically for now since this change does appear to alter the registration slightly but could consider doing so in the future.