Activate a Conda environment before attempting to launch application if necessary
The supplied setup scripts activate either Conda or Venv virtual environments in which the application typically must be run. In the main run script, check whether one of these environments is active before attempting to launch the GUI. If not, attempt to activate the Conda environment that setup script creates.

- Conda checker is based on https://stackoverflow.com/a/47610844/1911852
- Venv/virtualenv checker is based on https://stackoverflow.com/a/42580137/1911852
- Script directory locator is based on https://stackoverflow.com/a/3430395/1911852