Add `openjdk` as a dependency in Conda spec
Both Javabridge and ImageJ required Java. Setup has assumed that Java has been installed and set up correctly, but often Java has not been installed, added to the path, or installed as the required version (version 8 for ImageJ). To simplify installation, add the `openjdk` Conda package, pinning it at version 8. With Java 8 assumed to be installed, Javabridge could probably be changed to the official latest release (1.0.18), but we will keep it as the repo download since it continues to work, and the latest release still requires a compiler.