Remove Git-based install of Javabridge to remove Git requirement for Conda installs
The Git-based install has been necessary to support Java 12+, but the Conda-based setup now installs Java 8. To remove the requirement for Git to be present for Conda-based installs, remove this install. Javabridge will still be installed but through the installation of Python-Bioformats. The Git-based install remains in `requirements.txt` since Java is not installed in that pathway.