Add dedicated shutdown function
The application has typically been shutdown in three ways: 1) GUI window closure, which triggers the JVM and main database to be closed, 2) CLI finishing, which simply exits the main CLI function, and 3) early CLI exit, where the low-level `os._exit` function was called. Methods 2 and 3 performed no cleanup, and method 3 results in an error on Windows platforms because the code used is not supported there.

- Move shutdown tasks to a simple dedicated function including JVM closure and database shutdown
- Use `sys.exit` rather than `os._exit`, which prevents the exit code error on Windows and more generally performs other default cleanup tasks
- By closing the JVM, `os._exit` is no longer required to perform the exit