Specify ROC profiles to run during grid searches
Previously the `--roc` command line settings has been True/False, and all groups of settings in `roc_dict` have been run sequentially. Change the command line setting to take the names of the ROC groups to perform grid searches, which allows one or more groups to be run sequentially and the rest of the groups to be ignored. This change allows multiple grid search profiles to be set in `roc_dict` without having to run all of them together.

- Change `--roc` command line parameter from boolean to accept multiple strings specifying the names of grid search hyperparameter groups
- Change the current group name to "test"
- Add a "size" group setting to test a range of min/max sizes
- Fix Pandas errors when performing grid searches with different numbers of hyperparameters
- Remove unnecessary variables or initializations