Increase denoise ROI size in low resolution profile
Setting `denoise_size` to be the same as `segment_size` still led to edge artifacts because the segment ROI includes additional overlap regions. To ensure that the denoise ROI is simply the same as the segment ROI, increase the denoise size to be dramatically larger since it will be truncated to the max ROI shape.