Add theming capability to the graphical interface
Later PyQt versions appear to incorporate system theming into the interface, such as a dark theme in the ROI Selector GUI on MacOS with a dark system theme. The Matplotlib windows do not appear to incorporate this dark theme by default, however. A workaround is to use the "dark_background" Matplotlib theme, but it makes many other changes such as white borders around axes that may not fit the GUI.

To add customizability, add a theme setting as a command line parameter. Similar to profiles, this theme setting can be layered with multiple themes added, and each theme applies a set of RC parameters for Matplotlib. The RC parameters are stored in a `Themes` Enum class in `config`, with the `rc_params` dictionary moved there and the attribute changed to a list of currently set theme enums.

Currently I have not found a way to apply an RC setting for widget "colors," which are intensity values for buttons, text boxes, etc. Add a `widget_color` attribute with a default value set to the same as the default button color value. The custom style and theme setter function updates this value when applying a dark theme. Currently only the Atlas Editor uses these widgets, and the editor integrates the color value into its enable function.

Now let's hope that the energy spent supporting dark theming will be recovered by the display energy saved!