Fix activated Conda environment detection
The run script has checked for Conda environment activation by looking for the presence of a `conda-meta` file, but this file only indicates that the environment exists, not whether it is active. According to the Conda GitHub site, the `CONDA_PREFIX` environment is guaranteed to hold the path to the active environment (https://github.com/conda/conda/issues/3090). For simplicity, we will simply check for the presence of this variable to check for an active environment, which allows for the flexibility of environments that may not exactly match the given environment name.