Rename run script to include `.py` extension to allow launching by double-click
While the run script will run fine by double-click on OSes that identify Python by the shebang, Windows will not recognize the file without an extension. Add this extension to improve the chances that any Python executable set to run `.py` files will in fact run this script by double-click. This change also clarifies that the script is indeed a Python script.