Workaround for updating axes images where the first image has only one value
Atlases where the first labels plane shown contains only a single value do not update when the next images have more values, for unclear reasons. Simply generated a second value in the first image avoids this issue. As a workaround, check for single value arrays when showing the first labels image and increment the final value if necessary.