Option to treat a registered file prefix as a directory
The registered file writing `prefix` parameter has been assumed to be a file, but at times it may be preferable to give a directory to write files without the filename in `prefix`. This function uses `os.path.dirname` to find the directory of `prefix`, but if `prefix` is a directory, `os.path.dirname` will return the directory of this directory. Although a trailing file separator can be added, we will make the choice explicit by adding a separate parameter to flag treating `prefix` as a directory.