Fix vmin/vmax settings for overlaid images
Multichannel image display has expected a sequence for vmin/vmax arguments to apply to each channel. The global vmin/vmax values used in the ROI Editor overview plots are lists and worked directly in this image display. Commit eebe2831 (and more recently in commit 83e74eae) wrapped the values in sequences while introducing atlas label image overlays since the overlays were given as scalars, causing a regression in viewing overview plots when the global vmin/vmax values were applied. Commit 6bf1ef0e attempted to fix this regression but caused another regression in showing atlas label overlays.

Provide a more general fix for displaying overlaid images by allowing vmin/vmax settings as scalars or sequences. Numpy library functions frequently allow either scalars or "array-like" arguments, applying the same scalar to all values of an input array if a scalar is given. To support this functionality, add a library function that extracts a given element from a sequence if given or simply returns the value if not a sequence. This function also falls back to a default value is the given value is a sequence, but the index is not within range.

Use this function to accept scalars or sequences for vmin/vmax settings in image overlay display. This change avoids having to wrap scalars in sequences and allows the current settings to be used as-is for ROI Editor overview plots. Also apply the library function to alpha value extraction to ensure that the alpha value for the given channel is within range.