Incorporate the radius when determining whether to respond to a `DraggableCircle` press event
`DraggableCircle` objects respond to press events so long as the event occurs with a given modifier, within the circle's axes, and specifically within the circle itself. The `contains` function to check whether the event is within the circle appears to use a radius that may be too small, especially for small circles. Explicitly specify the circle's size so that the event can occur presumably within the full space of the circle.