Change truth database to check the given path first
Truth database loading has assumed that this database ends with a given suffix and is located in the current directory. Since the user is more likely to expect that an explicitly specified path is the full desired path, try this path first when loading a truth database. If not found, fall back to the prior method of adding the suffix if necessary and checking in the current directory.