Simplify the "contrast" microscope profile modifier to use the same colormap for all channels
Similar to the "bone" profile, simply use the same profile so that all channels will display in a high contrast colormap. The drawback is that all channels will appear the same, which can be addressed by adding more profiles with more customized settings or by making colormaps a command-line parameter.