Decouple experiment name parsing from truth DB loading during verification
During blob detection verification against truth sets, truth ROIs are loaded based on the experiment name. The experiment name had been parsed from the truth database name if the database was already loaded, or from the image filename if the truth database needed to be loaded. This method assumed that the experiment name in the database depended on how it was loaded, unnecessarily coupling the two processes. Instead, the experiment name should be assumed to be coupled to the current image name since presumably the automated detections are for the current image.

To decouple these processes and simplify loading, assume that the experiment name is based on the current image name. If no ROI is found matching this experiment, construct an ROI-based experiment name and check for ROIs. This approach supports images given as either a full image with offset/size parameters or a saved ROI-based image. If ROIs still cannot be found, raise an error and give a warning to avoid an exception.