Show truth blobs in verified plots with the blobs' given radius
Rather than showing truth blobs as a fixed size, use the blobs' actual size. The original purpose of the fixed size was to show the blobs as small circles, but this fixed size is not necessarily small for all resolutions. Using the blobs' actual size allows the blobs to scale with the given resolution and also to assess the truth set blob sizes.