Workaround for zoom or magnifications with negative values in overview plots
Overview plots indicate the total magnification by using the stored objective magnification and zoom values. The previous default values were -1.0, leading to potentially negative total magnification values. As a workaround, display the absolute value of the total magnification since it would always be expected to be > 0, assuming that this total magnification does not indicate inversion.