Change multichannel image display alpha default
Change the `alpha` parameter in `imshow_multichannel` to default to None for consistency with the `vmin` and `vmax` parameters in this function and `alphas` in `overlay_images`. Although alpha can be given to `imshow` as None directly, explicitly set it to 1 if None to allow further calculation based on the number of channels.