Allow padding a sequence with corresponding elements from another sequence
In the `libmag.pad_seq` function, change the `pad` parameter to accept a sequence as well. The missing values in `seq` will be filled with the corresponding values in `pad` if available. Otherwise, the last value in `pad` will be used to fill any remaining values. If `pad` is not a sequence, the `pad` value will continue to be used to fill `seq` repeatedly until reaching the desired length.