Remove series from filenames
The series string added to filenames was introduced to avoid naming conflicts when importing a multi-tile image. The series conflict is typically not encountered however since these multi-tile images are usually imported to a single stitched image. Since the series string is not used and requires additional path manipulation for each file, simply remove this string. The filename builder functions will continue to accept the `series` parameter since it may have alternative uses such as the multiple timepoints supported in the image5d format. Change the default series to None to ignore it for now.

While adjusting filenames, the suffixes for images and blobs have also been changed. Underscores are removed since they will be prepended in `libmag.combine_paths` and for consistency with registration suffix constants. The `.npz` in the image5d file is corrected to its `.npy` format. "Info" is changed to "meta" for clarity that it refers to a metadata file. "info_proc" is changed specifically to "blobs".