Avoid inserting string before non-existing extension
The `insert_before_ext` library function has attempted to insert a string before the extension of another string that lacks an extension by simply appending an empty extension. This workaround does not work for cases where the rightmost dot is in a parent directory since everything to the right of the dot include the filename are lost. Instead, only check for extensions within the basename. If no dot exists there, simply combine the string components.