Simplify modifying name for ROI-based image
Rather than re-appending the suffix, change the base filename to an ROI-based name if necessary and append the suffix to the final base filename. This filename can now be relied upon to refer to the actually loaded path. Also refactor the `image5d_proc` variable names to `roi` for clarity. The `suffix` parameter for `make_subimage_name` is no longer used but left here for now.