Distinguish sub-images from ROIs
MagellanMapper has supported regions of interest as a way to view or process subsets of an image. The `offset` and `roi_size` parameters have been used to define the boundaries of these ROIs and were later co-opted to define subimage boundaries. Sometimes both ROIs and sub-images need to be used together, however, such as ROIs within a sub-image. The command-line parameters have not distinguished between these two applications, which may cause confusion and unexpected behavior.

A workaround to distinguishing these two types of sub-images has been to save and load sub-images with filenames using offset/size parameters. This workaround has introduced new problems, however, such as the need to find the original image for metadata. The mechanism included loading a sub-image from a full image, which made ROI coordinates ambiguous as to whether they referred to the whole or sub-image.

As a first step to resolving these issues, we introduce here sub-image specific offset and size command-line argument to explicitly specify a sub-image. The ROI offset/size parameters have been refactored to distinguish them as ROI-related. Similar to these parameters, the sub-image offset/size parameters take coordinates in x,y,z order, but internally they switch the order to z,y,x for consistency with the processing within the software, and the ROI parameters are planned to be refactored to do the same eventually.

`setup_images` and `detect_blobs_large_image` now both expect to be given sub-image rather than ROI offset/size parameters. `prepare_roi` has been renamed to `prepare_subimg`, expecting size/offset in z,y,x and simplified under the assumption that only 5D images (or 4D if no channel) are given, and a new `prepare_roi` function serves as a wrapper to `prepare_subimg` but for x,y,z ordering. The `SUFFIX_ROI` has also been refactored to specify a sub-image.