Convert experiment names to sub-images when accessing ROIs from the database
Ground truth and verifications have assumed sub-image-based names. Now that sub-images are specified using the original filename plus separate parameters, update the experiment name used to save and retrieve experiment ROIs in the database. This method serves as a workaround to support previously saved filenames but may in the future be replaced with sub-image specific parameters in the database itself.