Fix importing image non-TIFF images
After changing the default `series` setting to None (commit 143e58e9), an error occurred during non-TIFF image import when trying to access the shape for the given series. Use a default series value of 0 in `read_file` if the argument is None. Since the series parameter is no longer needed, make it optional as well.