Remove loading blobs metadata
Blobs and sub-image metadata were saved in the same file, including information such as resolutions to allow using this metadata in place of the original image's metadata. The blobs and sub-image metadata did not provide sufficient information, however, leading to different results during detections compared with loading the original metadata. The blobs and sub-image metadata also stored information such as the basename in an attempt to reconstruct the path to the original metadata, but sub-images specified using the sub-image based name did not record the original image's filename, preventing the original file from being found.

Now that the original filename is given when loading sub-images, the original metadata can be accessed directly, precluding the need to load an alternate metadata file. To avoid loading this--potentially incomplete--metadata from the blobs and sub-image metadata file, only load the blobs and no other metadata from it. Also note that the additional metadata may no longer be stored in a future version. Refactor the out file variable names for clarity and consistency with the filename variables.