Crop full image when loading as sub-image
Sub-image loading will attempt to load a saved sub-image. If the file cannot be loaded, the fallback is to load the full image. Since the file was set to load as a sub-image, crop this full image to the corresponding sub-image offset/size.

Also group sub-image loading flag and feedback and clarify metadata loading. Distinguish sub-image and ROI preparation in feedback.