Option to load blobs without checking scaling
The blob loader can check for scaling with another image, but this scaling may not always be necessary. Add a parameter to control whether to check scaling and turn off the check by default. Make this check explicit since it will cause the register profile to be checked as well and alter the return signature.