Support loading sub-image based blobs independently of sub-images
Loading blobs for a sub-image has been dependent on loading the sub-image itself on the assumption that the blob coordinates need to be relative to the loaded image. Now that sub-images can be loaded from full images if a sub-image is specified but not available, blobs should similarly be loaded from either full archives or sub-image based archives.

If a sub-image is specified, first attempt to load blobs from the corresponding sub-image based blobs archive. Fall back to loading blobs from the full archive, cropping the blobs to those within the sub-image only. If no sub-image is specified, continue to load only the full blobs archive.