Use environment `JAVA_HOME` variable by default in stitch script
Default to using the environment variable since it will be set in the Conda (recommended) installation pathway and is generally recommended to be set. The pipeline script will only pass a Java home argument if set to avoid overwriting this variable in the stitch script. Also access the stitch type array using curly brackets for style consistency.