Support whole image blob detection with multiprocessing by spawning
Currently the Python multiprocessing library only supports starting processes by the spawn method on Windows, and the default method on Mac for Python 3.8 is now spawn as well. Unlike the fork method, spawn does not create a copy of the current environment, but rather starts a fresh environment. We have handled multiprocessing of large arrays but storing them as class attributes, but these attributes are not accessible in the newly spawned processes.

To support the spawn method, pass these class attributes to each process when in spawn mode. To avoid passing the entire large image, reload the image instead. For blob pruning, pickle each blob subset. Since performance is worsened with spawn compared with fork, continue to support the fork method.