Customize the multiprocessing start method
The Python `multiprocessing` library sets a default method by platform. Although only one method, "spawn", is available on Windows, multiple methods are available on Linux and MacOS. Moreover, the default method changed from "fork" to "spawn" on MacOS, even though the "fork" method is still available there. At times one method may be preferable over the default method, such as using the "fork" method on MacOS to take advantage of shared state even if there is a theoretical risk of state inconsistency is some cases (see https://bugs.python.org/issue33725).

To allow customization of the start method, add a new process settings profile option. Move multiprocessing-related library functions to the `chunking` module and add a method to set the start method, using this profile setting by default. To ensure that an invalid method is not applied, check that the chosen method is available on the current platform and fall back to the default method if the chosen method is not available.

Although the API doc for `set_start_method` says that the call should be protected in the name-main block, MagellanMapper does not have a single name-main block that is guaranteed to be called since there are multiple software entry points. Instead, the `cli.main` function is assumed to always be called to configure the software. The start method function also gives an exception when the method has already been set, allowing the error to be caught. Set up the start method in the CLI main function, simply catching any error during multiprocessing method setup.