Add troubleshooting doc on no valid OpenGL
In Windows Subsystem for Linux with a MobaXTerm X11 server, running the software in another terminal (ie Windows Terminal or the Ubuntu app rather than the MobaXTerm terminal) may lead to an inability to find a valid OpenGL implementation. A workaround is to use a LibGL or QT flag (see https://discourse.paraview.org/t/paraview-installation-error-in-linux-subsystem-windows-10/1381/3). During my testing, the LibGL flag fixed this issue.