Shift offset after cropping based on labels
Cropping the intensity image based on the extent of labels may cause the final out put image to be smaller. Relative to the original image, the resulting image's origin shifts by the start of cropping in each dimension. Update the final origin by this shift. Also clarify in the docstring for `replace_stick_with_numpy` that the SimpleITK Image object does not itself get changed, but a new Image object based on the given object has its array replaced.