Use an active flag to control the status of label mirroring
Label mirroring has used a separate profile option to turn it on or off. Following the active flag in the label edge setting, use this same type of flag in the mirror setting to group mirror settings together. The mirror scale setting is replaced with a dictionary to hold this flag and a `start` value to hold the current mirror setting, with the capacity for adding additional settings.

For now, we will keep the original `extend_labels` option until all options in it can be fully removed.