Option to turn off negative labeling
The inverse of labels have been applied to the opposite hemisphere from the labeled side to distinguish new from original labels. In some cases, the original labeling scheme may need to be applied to this opposite hemisphere to retain the original set of labels for both sides. To provide this option, add a new profile setting, `neg_labels`, to output only non-negative labels. Change the data type for these unsigned values to maximize the available range and potentially to decrease the required bit size. Also add a "noneg" profile to apply this option.