Change mirroring to previous behavior
Previously, label mirroring found the mirror index even if mirroring were turned off by the `extend_labels` profile setting, so long as the `mirror` setting itself were not None. To revert to this behavior, use the mirror `start` value even if the active flag is off. Default the start value to None again so mirroring is not applied even just for stats. Turn off the mirror flag for the CCFv3 profile so that it will not mirror but still provides the mirror start parameter to determine where to start making labels negative.