Make label edge and mirroring settings more consistent
The label edge and mirroring profile settings need a mechanism to turn them off while retaining customized parameters to generate metrics and find cropping boundaries even without including extended or mirrored labels in the final atlas. Previously the settings had been turned off through a combination of the now removed `extend_labels` setting, setting the `labels_mirror` or `labels_edge` setting groups to None, or setting the `start` parameter to None. This number of options made the behavior complex and inconsistent.

The recently added `ACTIVE` flag allows the settings to be turned off while keeping these parameters and replaced the `extend_labels` setting. Setting the parameter to None provided a convenient way to turn off the setting group when it was used for neither the output labels nor metrics/cropping, but it ran the risk of inadvertently re-setting the group to only a dictionary with only a subset of settings in a profile modifier. To avoid this issue, avoid setting the groups to None and instead use a `start` value of None to turn off the group. Change the default `labels_edge` `start` value to None so that edge extension is completely off by default.

This change also allows the `_curate_labels` function to avoid checking the `ACTIVE` flag. Instead, the function presumes that the given `mirror` or `edge` value will be used as is, ignoring the `ACTIVE` flag consistently for both values.